requirements.txt
slack-bolt
fastapi
uvicorn
python-dotenv
openai
tenacity
